requirements.txt
mkmsdk
pyyaml
